requirements.txt
<<<<<<< HEAD
argon2-cffi==20.1.0
asgiref==3.3.4
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
bleach==3.3.0
cffi==1.14.5
cycler==0.10.0
decorator==5.0.7
defusedxml==0.7.1
Django==3.2.2
django-rename-app==0.1.3
djangorestframework==3.12.4
entrypoints==0.3
et-xmlfile==1.0.1
future==0.18.2
ipykernel==5.5.3
ipython==7.23.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
iso8601==0.1.14
jedi==0.18.0
Jinja2==2.11.3
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.12
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kiwisolver==1.3.1
MarkupSafe==1.1.1
matplotlib==3.4.2
matplotlib-inline==0.1.2
mistune==0.8.4
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.3.0
numpy==1.20.2
openpyxl==3.0.7
packaging==20.9
pandocfilters==1.4.3
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
prometheus-client==0.10.1
prompt-toolkit==3.0.18
ptyprocess==0.7.0
pycparser==2.20
Pygments==2.8.1
pyparsing==2.4.7
pyrsistent==0.17.3
pyserial==3.5
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
pyzmq==22.0.3
qtconsole==5.0.3
QtPy==1.9.0
Send2Trash==1.5.0
six==1.15.0
sqlparse==0.4.1
stopwatch.py==1.0.1
terminado==0.9.4
testpath==0.4.4
tornado==6.1
traitlets==5.0.5
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
dash==1.13.3
=======
dash==1.13.3
pandas==1.0.5
gunicorn==20.0.4
>>>>>>> aa39903e7e0aee7ec4656021298c639e6bcfdd0b
